kms.library: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@39313 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/kms/kms_library.c
@@ -150,8 +150,8 @@ static ULONG KMS_Init(struct kms_base *KMSBase)
     if (error)
 	return FALSE;
 
-    KMSBase->rom_MapRawKey = SetFunction(KeymapBase, -7 * LIB_VECTSIZE, AROS_SLIB_ENTRY(patch_MapRawKey, Kms));
-    KMSBase->rom_MapANSI   = SetFunction(KeymapBase, -7 * LIB_VECTSIZE, AROS_SLIB_ENTRY(patch_MapRawKey, Kms));
+    KMSBase->rom_MapRawKey = SetFunction(KeymapBase, (LONG)(-7 * LIB_VECTSIZE), AROS_SLIB_ENTRY(patch_MapRawKey, Kms));
+    KMSBase->rom_MapANSI   = SetFunction(KeymapBase, (LONG)(-7 * LIB_VECTSIZE), AROS_SLIB_ENTRY(patch_MapRawKey, Kms));
 
     KMSBase->pub.kms_SwitchQual = KMS_QUAL_DISABLE;
